Displaying a very light placeholder bar for props with no votes (#343)
* Merge

* Placehoder vote bar when no votes

* Bg color

---------
