+++ src/components/Proposals/ProposalPage/VotingTimelineChart/VotingTimelineChart.tsx
@@ -2,27 +2,13 @@
 
 import Image from "next/image";
 import { useState } from "react";
-import {
-  LineChart,
-  Line,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  ResponsiveContainer,
-  ReferenceLine,
-  AreaChart,
-  Area,
-} from "recharts";
+import { Area, AreaChart, CartesianGrid, ReferenceLine, ResponsiveContainer, XAxis, YAxis } from "recharts";
 import { format } from "date-fns";
 import { icons } from "@/icons/icons";
 import { Proposal } from "@/app/api/common/proposals/proposal";
 import { Vote } from "@/app/api/common/votes/vote";
 import Tenant from "@/lib/tenant/tenant";
-import {
-  formatNumber,
-  formatNumberWithScientificNotation,
-  isScientificNotation,
-} from "@/lib/utils";
+import { formatNumber, formatNumberWithScientificNotation, isScientificNotation } from "@/lib/utils";
 
 const { token } = Tenant.current();
